Do not expose the token to the client
