+++ lib/azure_blob/auth.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "auth/msi_token_provider"
-
 module AzureBlob
   module Auth
   end
